data_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
python-binance==1.0.19
ccxt>=3.0.0
yfinance==0.2.31
ta==0.10.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
dask==2023.10.1
pydantic==2.4.2
python-dotenv==1.0.0
